Split off more functions into an internal directory.
